+++ bedhost/routers/api.py
@@ -2,7 +2,7 @@ import shlex
 import subprocess
 from typing import Optional
 
-from fastapi import APIRouter, HTTPException, Path, Query
+from fastapi import APIRouter, HTTPException, Path, Query, Response
 from fastapi.responses import PlainTextResponse, StreamingResponse
 
 from ..const import *
@@ -12,6 +12,13 @@ from ..main import _LOGGER, app, bbc
 import enum
 
 router = APIRouter()
+
+RemoteClassEnum = Enum(
+    "RemoteClassEnum",
+    {r: r for r in bbc.config[CFG_REMOTE_KEY]}
+    if is_data_remote(bbc)
+    else {"http": "http"},
+)
 
 FileColumnBedset = enum.Enum(
     "FileColumnBedset", get_enum_map(bbc, BEDSET_TABLE, "file")
@@ -30,7 +37,6 @@ file_map_bed = get_id_map(bbc, BED_TABLE, "file")
 img_map_bedset = get_id_map(bbc, BEDSET_TABLE, "image")
 
 img_map_bed = get_id_map(bbc, BED_TABLE, "image")
-
 
 ex_bed_digest = serve_columns_for_table(
     bbc=bbc, table_name=BED_TABLE, columns="md5sum", limit=1
@@ -80,6 +86,15 @@ async def get_version_info():
 
 
 # bed endpoints
+@router.get("/bed/genomes")
+async def get_bed_genome_assemblies():
+    """
+    Returns available genome assemblies in the database
+    """
+
+    return bbc.select_unique(table_name=BED_TABLE, column="genome")
+
+
 @router.get("/bed/all/data/count", response_model=int)
 async def get_bedfile_count():
     """
@@ -135,15 +150,45 @@ async def get_file_for_bedfile(
         condition_val=[md5sum],
         columns=["name", file_map_bed[id.value]],
     )[0][1]
-    remote = True if bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY] else False
+    remote = is_data_remote(bbc)
+
     path = (
-        os.path.join(bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY], file["path"])
+        os.path.join(bbc.config[CFG_REMOTE_KEY]["http"]["prefix"], file["path"])
         if remote
         else os.path.join(
             bbc.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY], file["path"]
         )
     )
+
     return serve_file(path, remote)
+
+
+@router.get("/bed/{md5sum}/file_path/{id}")
+async def get_file_for_bedfile(
+    md5sum: str = bd,
+    id: FileColumnBed = Path(..., description="File identifier"),
+    remoteClass: RemoteClassEnum = Query(
+        "http", description="Remote data provider class"
+    ),
+):
+    file = bbc.bed.select(
+        condition="md5sum=%s",
+        condition_val=[md5sum],
+        columns=["name", file_map_bed[id.value]],
+    )[0][1]
+    remote = is_data_remote(bbc)
+
+    path = (
+        os.path.join(
+            bbc.config[CFG_REMOTE_KEY][remoteClass.value]["prefix"], file["path"]
+        )
+        if remote
+        else os.path.join(
+            bbc.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY], file["path"]
+        )
+    )
+
+    return Response(path, media_type="text/plain")
 
 
 @router.get("/bed/{md5sum}/img/{id}")
@@ -161,10 +206,10 @@ async def get_image_for_bedfile(
         columns=["name", img_map_bed[id.value]],
     )[0][1]
 
-    remote = True if bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY] else False
+    remote = is_data_remote(bbc)
     path = (
         os.path.join(
-            bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY],
+            bbc.config[CFG_REMOTE_KEY]["http"]["prefix"],
             img["path" if format == "pdf" else "thumbnail_path"],
         )
         if remote
@@ -174,6 +219,39 @@ async def get_image_for_bedfile(
         )
     )
     return serve_file(path, remote)
+
+
+@router.get("/bed/{md5sum}/img_path/{id}")
+async def get_image_for_bedfile(
+    md5sum: str = bd,
+    id: ImgColumnBed = Path(..., description="Figure identifier"),
+    format: FigFormat = Query("pdf", description="Figure file format"),
+    remoteClass: RemoteClassEnum = Query(
+        "http", description="Remote data provider class"
+    ),
+):
+    """
+    Returns the bedfile plot with provided ID in provided format
+    """
+    img = bbc.bed.select(
+        condition="md5sum=%s",
+        condition_val=[md5sum],
+        columns=["name", img_map_bed[id.value]],
+    )[0][1]
+
+    remote = is_data_remote(bbc)
+    path = (
+        os.path.join(
+            bbc.config[CFG_REMOTE_KEY][remoteClass.value]["prefix"],
+            img["path" if format == "pdf" else "thumbnail_path"],
+        )
+        if remote
+        else os.path.join(
+            bbc.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY],
+            img["path" if format == "pdf" else "thumbnail_path"],
+        )
+    )
+    return Response(path, media_type="text/plain")
 
 
 @router.get("/bed/{md5sum}/regions/{chr_num}", response_class=PlainTextResponse)
@@ -193,9 +271,9 @@ def get_regions_for_bedfile(
         columns=["name", "bigbedfile"],
     )[0][1]
 
-    remote = True if bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY] else False
+    remote = is_data_remote(bbc)
     path = (
-        os.path.join(bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY], file["path"])
+        os.path.join(bbc.config[CFG_REMOTE_KEY]["http"]["prefix"], file["path"])
         if remote
         else os.path.join(
             bbc.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY], file["path"]
@@ -236,6 +314,13 @@ def get_regions_for_bedfile(
 
 
 # bedset endpoints
+@router.get("/bedset/genomes")
+async def get_bedset_genome_assemblies():
+    """
+    Returns available genome assemblies in the database
+    """
+
+    return bbc.select_unique(table_name=BEDSET_TABLE, column="genome")
 
 
 @router.get("/bedset/all/data/count", response_model=int)
@@ -315,15 +400,43 @@ async def get_file_for_bedset(
         condition_val=[md5sum],
         columns=["name", file_map_bedset[id.value]],
     )[0][1]
-    remote = True if bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY] else False
+    remote = is_data_remote(bbc)
     path = (
-        os.path.join(bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY], file["path"])
+        os.path.join(bbc.config[CFG_REMOTE_KEY]["http"]["prefix"], file["path"])
         if remote
         else os.path.join(
             bbc.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY], file["path"]
         )
     )
+
     return serve_file(path, remote)
+
+
+@router.get("/bedset/{md5sum}/file_path/{id}")
+async def get_file_path_for_bedset(
+    md5sum: str = bsd,
+    id: FileColumnBedset = Path(..., description="File identifier"),
+    remoteClass: RemoteClassEnum = Query(
+        "http", description="Remote data provider class"
+    ),
+):
+    file = bbc.bedset.select(
+        condition="md5sum=%s",
+        condition_val=[md5sum],
+        columns=["name", file_map_bedset[id.value]],
+    )[0][1]
+    remote = is_data_remote(bbc)
+    path = (
+        os.path.join(
+            bbc.config[CFG_REMOTE_KEY][remoteClass.value]["prefix"], file["path"]
+        )
+        if remote
+        else os.path.join(
+            bbc.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY], file["path"]
+        )
+    )
+
+    return Response(path, media_type="text/plain")
 
 
 @router.get("/bedset/{md5sum}/img/{id}")
@@ -331,6 +444,9 @@ async def get_image_for_bedset(
     md5sum: str = bsd,
     id: ImgColumnBedset = Path(..., description="Figure identifier"),
     format: FigFormat = Query("pdf", description="Figure file format"),
+    remoteClass: RemoteClassEnum = Query(
+        "http", description="Remote data provider class"
+    ),
 ):
     """
     Returns the img with provided ID
@@ -341,10 +457,10 @@ async def get_image_for_bedset(
         columns=["name", img_map_bedset[id.value]],
     )[0][1]
 
-    remote = True if bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY] else False
+    remote = is_data_remote(bbc)
     path = (
         os.path.join(
-            bbc.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY],
+            bbc.config[CFG_REMOTE_KEY]["http"]["prefix"],
             img["path" if format == "pdf" else "thumbnail_path"],
         )
         if remote
@@ -353,4 +469,39 @@ async def get_image_for_bedset(
             img["path" if format == "pdf" else "thumbnail_path"],
         )
     )
+
     return serve_file(path, remote)
+
+
+@router.get("/bedset/{md5sum}/img_path/{id}")
+async def get_image_for_bedset(
+    md5sum: str = bsd,
+    id: ImgColumnBedset = Path(..., description="Figure identifier"),
+    format: FigFormat = Query("pdf", description="Figure file format"),
+    remoteClass: RemoteClassEnum = Query(
+        "http", description="Remote data provider class"
+    ),
+):
+    """
+    Returns the img with provided ID
+    """
+    img = bbc.bedset.select(
+        condition="md5sum=%s",
+        condition_val=[md5sum],
+        columns=["name", img_map_bedset[id.value]],
+    )[0][1]
+
+    remote = is_data_remote(bbc)
+    path = (
+        os.path.join(
+            bbc.config[CFG_REMOTE_KEY][remoteClass.value]["prefix"],
+            img["path" if format == "pdf" else "thumbnail_path"],
+        )
+        if remote
+        else os.path.join(
+            bbc.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY],
+            img["path" if format == "pdf" else "thumbnail_path"],
+        )
+    )
+
+    return Response(path, media_type="text/plain")