+++ bedhost/routers/private_api.py
@@ -10,6 +10,29 @@ from ..main import _LOGGER, app, bbc
 router = APIRouter()
 
 # private API
+@router.get("/distance/{term}/bedfiles/{genome}", response_model=DBResponse)
+async def get_bedfiles_in_distance(
+    term: str = Path(..., description="search term"),
+    genome: str = Path(..., description="genome assemblies"),
+    ids: Optional[List[str]] = Query(None, description="Bedfiles table column name"),
+    limit: int = Query(None, description="number of rows returned by the query"),
+):
+    term = term.replace(" ", ",").split(",")
+
+    if ids:
+        assert_table_columns_match(bbc=bbc, table_name=BED_TABLE, columns=ids)
+    res = bbc.select_bedfiles_for_distance(
+        genome=genome, condition_val=term, bedfile_col=ids, limit=limit
+    )
+    if res:
+        colnames = list(res[0].keys())
+        values = [list(x.values()) for x in res]
+        _LOGGER.info(f"Serving data for columns: {colnames}")
+    else:
+        _LOGGER.warning("No records matched the query")
+        colnames = []
+        values = [[]]
+    return {"columns": colnames, "data": values}
 
 
 @router.get("/query/{table_name}/{query}", include_in_schema=False)
@@ -41,7 +64,7 @@ async def get_query_results(
         raise HTTPException(status_code=404, detail=msg)
 
 
-@router.get("/filters/{table_name}", include_in_schema=False)
+@router.get("/filters/{table_name}")
 async def get_search_setup_for_table(
     table_name: TableName = Path(..., description="DB Table name")
 ):