+++ docs/changelog.md
@@ -2,6 +2,13 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format. 
 
+## [0.0.6] -- 2021-05-17
+### add
+- Add endpoints that serve:
+  - a list of genome assemblies in bedsets and bedfiles table
+  - bed files by search term(s)
+  - remote file path (http / s3)
+  
 ## [0.0.5] -- 2021-04-15
 ### add
 - Add examples of API endpoints